+++ packages/cms/lib/modules/icon-section-widgets/index.js
@@ -60,6 +60,7 @@ module.exports = {
                     label: 'Description',
                     type: 'text',
                     type: 'string',
+                    textarea: true
                 },
                 {
                   name: 'image',