+++ CHANGELOG.md
@@ -7,6 +7,8 @@
 * Added possibility to query GET overview routes with NOT (!=) and substring (like%%) operator
 * Add option to list user data publically or per role
 * Allow viewablebyRole to be edited by the user
+* Open /stats/site/:SITE_ID/vote/total for everybody
+* Add sanitizeIfNotNull decorator to the sanitize util to prevent transforming of null into the string 'null'
 
 ## v0.8.0 (2020-11-02)
 * Add id & extraData to properties included in idea GET call with param includeUser=1