+++ CHANGELOG.md
@@ -1,12 +1,12 @@
 # Changelog
 
-
 ## Unreleased updates
+* Make description textarea in icon-section
+
+## v0.9.0 (Pre release)
 * Fix: Ideas Map had an option to not link Flags. The option had disappeared but is now back.
 * Feature: Better explanations in global newsletter settings
 * Add admin warning to vote counter if vote count is not publicly available
-
-## v0.9.0 (Pre release)
 * Update NPM modules for security
 * Don't add JWT when doing a resource overview call since results are cached
 * Empty cache on DELETE call via api proxy